+++ secrets/hash_test.go
@@ -4,4 +4,5 @@ import "testing"
 
 func TestHash(t *testing.T) {
 	// TODO: HashPassword
+	// TODO: MatchPassword
 }